requirements.txt
openai
aiogram
python-dotenv
